requirements.txt
absl-py==0.2.1
aniso8601==3.0.0
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gast==0.2.0
GeoAlchemy2==0.4.2
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.6
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.3
Pillow==5.1.0
protobuf==3.5.2.post1
psycopg2==2.7.4
pytz==2018.4
PyYAML==3.12
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.7
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
uwsgi